Modal components moved to components/modal emptyField comp created and implemented, some css change
