+++ client/src/Pages/Collection/Playlists.scss
@@ -6,13 +6,4 @@
         grid-template-columns: repeat( auto-fill, minmax(190px, 1fr));
         gap: 1.5rem;
     }        
-    
-
-    &__empty{
-        flex: 1;
-        display: flex;
-        align-items: center;
-        justify-content: center;
-        height: 100%;
-    }
 }
