+++ client/src/components/Modal/ModalCloseBtn.scss
