+++ client/src/components/Modal/ModalHeader.scss
