+++ client/src/modals/PasswordChangeModal.jsx
@@ -2,7 +2,7 @@ import React from 'react'
 import clsx from 'clsx'
 import CustomInput from '@/components/CustomInput'
 import PasswordInput from '@/components/PasswordInput'
-import ModalHeader from '@/modals/ModalHeader'
+import ModalHeader from '@/components/Modal/ModalHeader'
 import { updateUserPassword, updateMail } from '@/firebase/auth'
 import { passwordChangeSchema, emailChangeSchema } from '@/forms/schemas'
 import { auth } from '@/firebase/auth'
